[Download] Improved downloading utility stability

+++ illadownload/src/illarion/download/tasks/unpack/Unpack.java
@@ -1,67 +1,62 @@
 /*
- * This file is part of the Illarion Download Manager.
- * 
- * Copyright © 2011 - Illarion e.V.
- * 
- * The Illarion Download Manager is free software: you can redistribute i and/or
- * modify it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or (at your
- * option) any later version.
- * 
- * The Illarion Download Manager is distributed in the hope that it will be
- * useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License along with
- * the Illarion Download Manager. If not, see <http://www.gnu.org/licenses/>.
+ * This file is part of the Illarion Download Utility.
+ *
+ * Copyright © 2012 - Illarion e.V.
+ *
+ * The Illarion Download Utility is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * The Illarion Download Utility is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with the Illarion Download Utility.  If not, see <http://www.gnu.org/licenses/>.
  */
 package illarion.download.tasks.unpack;
 
-import java.io.BufferedInputStream;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
+import illarion.common.util.DirectoryManager;
+import illarion.common.util.Pack200Helper;
+import illarion.download.install.resources.ResourceManager;
+import illarion.download.tasks.download.Download;
+import illarion.download.tasks.download.DownloadResult;
+import org.tukaani.xz.XZInputStream;
+
+import java.io.*;
 import java.net.URL;
 import java.nio.channels.Channels;
 import java.nio.channels.FileChannel;
 import java.nio.channels.ReadableByteChannel;
 import java.util.ArrayList;
-import java.util.List;
+import java.util.Collection;
 import java.util.concurrent.Callable;
 import java.util.jar.JarOutputStream;
 import java.util.jar.Pack200;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
-import org.tukaani.xz.XZInputStream;
-
-import illarion.common.util.DirectoryManager;
-import illarion.common.util.Pack200Helper;
-
-import illarion.download.install.resources.ResourceManager;
-import illarion.download.tasks.download.Download;
-import illarion.download.tasks.download.DownloadResult;
-
 /**
  * This class defines the task that actually takes care for unpacking the
  * resources to the hard drive.
- * 
+ *
  * @author Martin Karing
- * @since 1.00
  * @version 1.01
+ * @since 1.00
  */
 public final class Unpack implements Callable<UnpackResult> {
     /**
      * The monitor that is applied to the monitor input stream that is used to
      * generate the callback to the other parts of the application.
-     * 
+     *
      * @author Martin Karing
-     * @since 1.00
      * @version 1.00
+     * @since 1.00
      */
     private static final class StreamMonitor implements
-        CountBytesInputStream.Callback {
+            CountBytesInputStream.Callback {
         /**
          * The full file size that is reported to the manager.
          */
@@ -78,16 +73,15 @@ public final class Unpack implements Callable<UnpackResult> {
         private final UnpackManager unpackManager;
 
         /**
-         * Create a new stream monitor and setup all required informations to
-         * get it working properly.
-         * 
-         * @param upManager the manager that will be informed about the process
-         * @param size the total amount of data that is reported to the manager
+         * Create a new stream monitor and setup all required information to get it working properly.
+         *
+         * @param upManager    the manager that will be informed about the process
+         * @param size         the total amount of data that is reported to the manager
          * @param parentUnpack the unpack task that is the parent to this
-         *            monitor
+         *                     monitor
          */
-        public StreamMonitor(final UnpackManager upManager, final long size,
-            final Unpack parentUnpack) {
+        private StreamMonitor(final UnpackManager upManager, final long size,
+                              final Unpack parentUnpack) {
             unpackManager = upManager;
             fileSize = size;
             parent = parentUnpack;
@@ -128,24 +122,24 @@ public final class Unpack implements Callable<UnpackResult> {
     /**
      * Create the unpack task based on a download that was handled before. This
      * unpacking task will try to extract the data downloaded.
-     * 
-     * @param download the download that is further processed by this task
-     * @param result the result of the download
+     *
+     * @param download   the download that is further processed by this task
+     * @param result     the result of the download
      * @param unpManager the unpacking manager that handles this task
      */
     Unpack(final Download download, final DownloadResult result,
-        final UnpackManager unpManager) {
+           final UnpackManager unpManager) {
         file = download.getTarget();
         targetDir =
-            new File(DirectoryManager.getInstance().getDataDirectory(),
-                download.getDirectory());
+                new File(DirectoryManager.getInstance().getDataDirectory(),
+                        download.getDirectory());
         downloadResult = result;
         manager = unpManager;
         name = download.getName();
     }
 
     @Override
-    public UnpackResult call() throws Exception {
+    public UnpackResult call() throws IOException {
         final UnpackResult result = callImpl();
         manager.reportFinished(this, result);
         return result;
@@ -154,54 +148,52 @@ public final class Unpack implements Callable<UnpackResult> {
     /**
      * This is the implementation of the call function that is launched when
      * this task is executed.
-     * 
+     *
      * @return the result of the unpacking operation
-     * @throws Exception in case anything at all goes wrong
+     * @throws IOException in case unpacking the file fails
      */
-    public UnpackResult callImpl() throws Exception {
+    public UnpackResult callImpl() throws IOException {
         switch (downloadResult.getResult()) {
             case canceled:
                 return new UnpackResult(name, UnpackResult.Results.canceled,
-                    "unpack.cancled", file); //$NON-NLS-1$
+                        "unpack.cancled", file); //$NON-NLS-1$
             case downloadFailed:
                 return new UnpackResult(name, UnpackResult.Results.canceled,
-                    "unpack.downloadfailed", file); //$NON-NLS-1$
+                        "unpack.downloadfailed", file); //$NON-NLS-1$
             case notModified:
                 return new UnpackResult(name,
-                    UnpackResult.Results.notModified,
-                    "unpack.notmodified", file); //$NON-NLS-1$
+                        UnpackResult.Results.notModified,
+                        "unpack.notmodified", file); //$NON-NLS-1$
             case downloaded:
                 break;
         }
 
         if (!targetDir.exists() && !targetDir.mkdirs()) {
             throw new IllegalStateException(
-                "Can't create required directories."); //$NON-NLS-1$
+                    "Can't create required directories."); //$NON-NLS-1$
         }
 
-        CountBytesInputStream cIn = null;
         ZipInputStream zIn = null;
         final long fileSize = file.length();
         final long blockSize = Math.max(1024, fileSize / 1000);
 
-        final List<File> installedFiles = new ArrayList<File>();
+        final Collection<File> installedFiles = new ArrayList<File>();
 
         ReadableByteChannel inChannel = null;
         FileChannel outChannel = null;
         JarOutputStream jOutStream = null;
         try {
-            cIn = new CountBytesInputStream(new FileInputStream(file));
+            final CountBytesInputStream cIn = new CountBytesInputStream(new FileInputStream(file));
 
-            cIn.addCallback(new StreamMonitor(manager, fileSize, this));
+            cIn.addCallback(new Unpack.StreamMonitor(manager, fileSize, this));
             cIn.setCallbackInterval(blockSize);
 
             zIn =
-                new ZipInputStream(new BufferedInputStream(
-                    new XZInputStream(cIn)));
+                    new ZipInputStream(new BufferedInputStream(
+                            new XZInputStream(cIn)));
             inChannel = Channels.newChannel(zIn);
 
             ZipEntry currEntry = zIn.getNextEntry();
-            long posInFile;
             while (currEntry != null) {
                 final String entryName = currEntry.getName();
                 boolean pack200 = false;
@@ -211,10 +203,20 @@ public final class Unpack implements Callable<UnpackResult> {
                 final File targetFile = new File(targetDir, entryName);
                 final String fullPath = targetFile.getAbsolutePath();
 
-                new File(fullPath.substring(0, fullPath.lastIndexOf(File.separatorChar))).mkdirs();
+                final File targetDir = new File(fullPath.substring(0, fullPath.lastIndexOf(File.separatorChar)));
+                if (targetDir.exists() && !targetDir.isDirectory()) {
+                    if (!targetDir.delete()) {
+                        return new UnpackResult(name, UnpackResult.Results.failed,
+                                "unpack.createDirsFailed", file); //$NON-NLS-1$
+                    }
+                }
+                if (!targetDir.exists() && !targetDir.mkdirs()) {
+                    return new UnpackResult(name, UnpackResult.Results.failed,
+                            "unpack.createDirsFailed", file); //$NON-NLS-1$
+                }
 
                 outChannel = new FileOutputStream(targetFile).getChannel();
-                posInFile = 0;
+                long posInFile = 0;
                 while (zIn.available() > 0) {
                     posInFile += outChannel.transferFrom(inChannel, posInFile, fileSize);
                 }
@@ -223,15 +225,15 @@ public final class Unpack implements Callable<UnpackResult> {
 
                 if (pack200) {
                     final File targetFileP200 =
-                        new File(targetDir, entryName.substring(0,
-                            entryName.length() - 5));
+                            new File(targetDir, entryName.substring(0,
+                                    entryName.length() - 5));
                     final Pack200.Unpacker p200unpacker =
-                        Pack200Helper.getUnpacker();
+                            Pack200Helper.getUnpacker();
                     installedFiles.add(targetFileP200);
 
                     jOutStream =
-                        new JarOutputStream(new FileOutputStream(
-                            targetFileP200));
+                            new JarOutputStream(new FileOutputStream(
+                                    targetFileP200));
                     p200unpacker.unpack(targetFile, jOutStream);
 
                     jOutStream.finish();
@@ -277,7 +279,7 @@ public final class Unpack implements Callable<UnpackResult> {
 
     /**
      * Get the name of this unpacking task. This name is displayed in the GUI.
-     * 
+     *
      * @return the name of this task
      */
     public String getName() {
